Add files via upload
Record on keypress space key

+++ src/recording_listener.py
@@ -1,12 +1,13 @@
 import os
 import sys
 import time
-import main
 import logging
 import configparser
 import constants as const
 from watchdog.observers import Observer
 from watchdog.events import LoggingEventHandler
+from speech_to_text import SpeechToText
+from language_translator import LanguageTranslation
 
 configParser = configparser.ConfigParser()
 configParser.read(const.CONFIG_FILE_NAME)
@@ -15,15 +16,27 @@ audio_file_name = configParser.get(const.CFG_AUDIO_DIR, const.CFG_AUDIO_FILE_NAM
 
 class Event(LoggingEventHandler):
     def on_created(self, event):
+        audio_file_path = os.path.join(audio_path, audio_file_name)
         try:
             print("Triggering speech to text ...")
             time.sleep(1)
-            main.start()
-            print("Deleting the recording file ...")
-            audio_file_path = os.path.join(audio_path, audio_file_name)
-            os.remove(audio_file_path)
-        except:
+            start_time = time.time()
+            extracted_speech, baseLanguage = SpeechToText().start_speech_to_text()
+            if baseLanguage == "none":
+                print("Invalid Language")
+                return
+            else:
+                translated_text = LanguageTranslation().start_translation(extracted_speech, baseLanguage)
+                print(translated_text)
+                print("--- TOTAL execution time: %s seconds ---" % (time.time() - start_time))
+                print("Deleting the recording file ...")
+                os.remove(audio_file_path)
+        except Exception as e:
+            print(e)
             print("An error ocurred while translating speech to text.")
+        finally:
+            if os.path.isfile(audio_file_path):
+                os.remove(audio_file_path)
 
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO,
